imgur/requirements.txt
imgurpython==1.1.7
python-decouple==3.0
